Move font awesome init to utils.
